Applied Fix for #37
Changed GetFullPath('/', $false) to GetFullPath('/', $true), this mean
that the entire path will be returned and used for comparisons.  The
main difference is that the top level folder aka the database name must
be included in the the path.

Updated documentation for this change.

Updated Get-KPGroup To be Case Insensitive for path look up.