Simplified Core Functions Connection Building
Created function Get-KPDynamicParameters to build the database profile
name and master key parameters.

added masterkey dynamic parameter to allow passing a masterkey for
scripting. The old behaviour was to prompt for the masterkey. It will
still do this if one is required but not passed from  a core function.

created function Invoke-KPConnection to get a profile, credential object
and then finally connection.

these changes simplied and removed tons of redunant code.